# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing public demo sandboxes and internal employee sandboxes.",
        "title": "Shopshredder Sandbox API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/me/sandboxes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns sandboxes owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "List my sandboxes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Sandbox"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/public/demos": {
            "post": {
                "description": "Spin up a new sandbox for a guest visitor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Create a public demo sandbox",
                "parameters": [
                    {
                        "description": "Sandbox configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSandboxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Sandbox"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/public/sandboxes": {
            "get": {
                "description": "Returns sandboxes for the current guest session (cookie-based)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "List guest sandboxes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Sandbox"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/public/sandboxes/{id}": {
            "delete": {
                "description": "Stop and remove a guest session's sandbox",
                "tags": [
                    "Public"
                ],
                "summary": "Delete a guest sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sandboxes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all sandboxes that are currently active (admin view)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "List all active sandboxes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Sandbox"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Spin up a new sandbox for an authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "Create a private sandbox",
                "parameters": [
                    {
                        "description": "Sandbox configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSandboxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Sandbox"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sandboxes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single sandbox by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "Get sandbox by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Sandbox"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stop and remove an authenticated user's sandbox",
                "tags": [
                    "Sandboxes"
                ],
                "summary": "Delete a sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sandboxes/{id}/snapshot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Commit the current state of a running sandbox as a new Docker image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "Create a snapshot image from a sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Snapshot metadata",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSnapshotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Image"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sandboxes/{id}/ttl": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add additional time to a sandbox's expiration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sandboxes"
                ],
                "summary": "Extend sandbox TTL",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TTL extension",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExtendTTLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Sandbox"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateSandboxRequest": {
            "type": "object",
            "required": [
                "imageId"
            ],
            "properties": {
                "imageId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8ae13ed9-cfb1-4941-a248-bc74b9fb6a24"
                },
                "ttlMinutes": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateSnapshotRequest": {
            "type": "object",
            "required": [
                "name",
                "tag"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "isPublic": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "thumbnailUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                }
            }
        },
        "dto.ExtendTTLRequest": {
            "type": "object",
            "properties": {
                "ttlMinutes": {
                    "type": "integer"
                }
            }
        },
...
        "models.Sandbox": {
            "type": "object",
            "properties": {
                "clientIp": {
                    "type": "string"
                },
                "containerId": {
                    "type": "string"
                },
                "containerName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "guestSessionId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "status": {
                    "enum": [
                        "starting",
                        "running",
                        "stopped",
                        "expired",
                        "deleted",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SandboxStatus"
                        }
                    ],
                    "example": "running"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.SandboxStatus": {
            "type": "string",
            "enum": [
                "starting",
                "running",
                "stopped",
                "expired",
                "deleted",
                "failed"
            ],
            "x-enum-varnames": [
                "SandboxStatusStarting",
                "SandboxStatusRunning",
                "SandboxStatusStopped",
                "SandboxStatusExpired",
                "SandboxStatusDeleted",
                "SandboxStatusFailed"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}